Initialize gap buffer from a given string
The buffers internal vector was changed from storing characters
(UTF-32) to unsigned 8-bit integers (u8). Rust treats strings as
collections of UTF-8 bytes and allows easy conversion of strings
to vectors of 8-bit integers and vice versa. This seems to be more
in line with Rust convention.

At the moment it is only possible to add valid UTF-8 to the buffer
via a string and output valid UTF-8. Future feature additions will
allow for the insertion of single characters. Currently the buffer
"unwraps" the conversion of the internal vector to a string. Any
invalid UTF-8 bytes will cause a panic.