Name project 'One Editor' with new scope
The project is now called 'One Editor'.

The project started out as a gap buffer: a core component of a text
editor. The gap buffer is close to MPV for its features. It's time
to build a text editor around it to produce a working MVP product.

Building a text editor has always been a long term goal. The gap
buffer is at the 'good enough' stage. Instead of making the gap
buffer a separate library it will be a part of the text editor code
base.